requirements.txt
kafka-python==2.0.2
faker==13.3.0
